Switch to uint64_t seeds and std::MT19937_64 for host RNG.
This makes the full range of entropy available for curand, and upgrades host rng engines to the 64 bit version of mt19937

+++ include/flamegpu/runtime/utility/HostRandom.cuh
@@ -50,11 +50,11 @@ class HostRandom {
      * Change the seed used for random generation
      * @param seed New random seed
      */
-    void setSeed(const unsigned int &seed);
+    void setSeed(const uint64_t &seed);
     /**
      * Returns the last value used to seed random generation
      */
-    unsigned int getSeed() const;
+    uint64_t getSeed() const;
 
  private:
     explicit HostRandom(RandomManager &_rng) : rng(_rng) { }
