+++ src/flamegpu/sim/RunPlan.cpp
@@ -25,7 +25,9 @@ RunPlan& RunPlan::operator=(const RunPlan& other) {
         this->property_overrides.emplace(i.first, util::Any(i.second));
     return *this;
 }
-void RunPlan::setRandomSimulationSeed(const unsigned int &_random_seed) { random_seed = _random_seed; }
+void RunPlan::setRandomSimulationSeed(const uint64_t &_random_seed) {
+    random_seed = _random_seed;
+}
 void RunPlan::setSteps(const unsigned int &_steps) {
     if (_steps == 0 && !allow_0_steps) {
         throw std::out_of_range("Model description requires atleast 1 exit condition to have unlimited steps, "
@@ -37,7 +39,7 @@ void RunPlan::setOutputSubdirectory(const std::string &subdir) {
     output_subdirectory = subdir;
 }
 
-unsigned int RunPlan::getRandomSimulationSeed() const {
+uint64_t RunPlan::getRandomSimulationSeed() const {
     return random_seed;
 }
 unsigned int RunPlan::getSteps() const {
